+++ yggdrasil_decision_forests/port/python/ydf/learner/wrapper/wrapper_generator.cc
@@ -257,7 +257,7 @@ included for reference only. The actual wrappers are re-generated during
 compilation.
 """
 
-from typing import Dict, Optional
+from typing import Dict, Optional, Sequence
 $0
 
 )",
@@ -502,6 +502,9 @@ $2
     tuner: If set, automatically select the best hyperparameters using the
       provided tuner. When using distributed training, the tuning is
       distributed.
+    workers: If set, enable distributed training. "workers" is the list of IP
+      addresses of the workers. A worker is a process running
+      `ydf.start_worker(port)`.
   """
 
   def __init__(self,
@@ -523,6 +526,7 @@ $3,
       resume_training: bool = False,
       resume_training_snapshot_interval_seconds: int = 1800,
       tuner: Optional[tuner_lib.AbstractTuner] = None,
+      workers: Optional[Sequence[str]] = None,
       ):
 
     hyper_parameters = {
@@ -543,6 +547,7 @@ $4
         resume_training=resume_training,
         resume_training_snapshot_interval_seconds=resume_training_snapshot_interval_seconds,
         working_dir=working_dir,
+        workers=workers,
     )
 
     super().__init__(learner_name="$1",